Move hook to utility function to set user info after a sucessful login
